security on post pull delete

+++ pages/api/products/[id].js
@@ -4,8 +4,11 @@ import Product from '../../../models/Product'
 export default async function handler(req, res) {
     const {
         method,
+        cookies,
         query: { id }
     } = req;
+
+    const token = cookies.token;
 
     dbConnect()
 
@@ -19,6 +22,10 @@ export default async function handler(req, res) {
     }
 
     if (method == "PUT") {
+        if (!token || token != process.env.TOKEN) {
+            return res.status(401).json("Not Authenticated.")
+        }
+
         try {
             const product = await Product.create(req.body);
             res.status(201).json(product);
@@ -28,6 +35,10 @@ export default async function handler(req, res) {
     }
 
     if (method == "DELETE") {
+        if (!token || token != process.env.TOKEN) {
+            return res.status(401).json("Not Authenticated.")
+        }
+
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json("The product has been deleted.");
